portfolio-backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Portfolio Backend API - Authentication & Admin Service"
authors = [
    {name = "Eduardo Alves",email = "eduardo.alves@centranet.com.br"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.136.0,<0.137.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "redis[hiredis] (>=7.4.0,<8.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "cryptography (>=46.0.7,<47.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "prometheus-fastapi-instrumentator (>=7.0.0,<8.0.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "fakeredis[lua] (>=2.24,<3.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
# E/W/F = pyflakes + pycodestyle, I = isort, B = bugbear, UP = pyupgrade,
# S = flake8-bandit (security). SIM = flake8-simplify, RUF = Ruff-specific.
select = ["E", "W", "F", "I", "B", "UP", "S", "SIM", "RUF"]
ignore = [
    "E501",   # line-too-long — delegated to the formatter
    "B008",   # allow FastAPI's Depends() in default arguments
    "S101",   # allow `assert` in tests
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["S603", "S607"]   # scripts may invoke subprocesses
"tests/*"   = ["S105", "S106"]   # fixture passwords are acceptable in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --strict-markers"
